requirements.txt
caldav==0.5.0
certifi==2018.1.18
chardet==3.0.4
coverage==4.5.1
idna==2.6
lxml==4.2.1
nose==1.3.7
oauthlib==2.0.7
py-trello==0.10.0
python-dateutil==2.7.2
pytz==2018.4
PyYAML==3.12
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
urllib3==1.22
uuid==1.30
vobject==0.9.5
